+++ examples/dithering3_3/pass0/compute.js
