+++ examples/dithering3_3/pass0/frag.js
